Modifying the class ComplementTypes.scala
The current version of the file does not handle complement with
multiple languages. The important thing here is to let the user
choose all the languages he wants to complement his original
language.

Adds support to complement types in multiple languages in the core
routine.

Adds one more helper function to check the length of the number of
languages array.
   def testArrayLength(anArrayLength: Int, aLog: Log)

Adds four new arrays to hold the language names, the instance
types files names, the tdb store paths and file iterators
   allLangsArray of type String
   instTypesNamesArray of type String
   tdbStoreArray of type Model
   fileIteratorArray of type StmtIterator

Modifies a bit the code structure to be more concise