Modifying the ComplementTypes class
Modifies the stucture of the code to be cleaner.

Adds support to query Freebase endpoint for types.

Adds support to query Freebase locally for types.

Modifies the structure of the code to separate the methods inside
the class by purpose.

Adds three new classes to do the types complement task
    compTypesWithFreebaseRemote()
    compTypesWithFreebaseLocal()
    compTypesWithOtherLanguages()