Creating class ExtractLabels
Adds a new class to org.dbepedia.spotlight.util package in the index
module. The class is named ExtractLabels.scala.

Changes the location of the ComplementTypes class to the org.dbpedia.
spotlight.util package inside the index module.

Adds one entry to the indexing.properties file.

Modifies the ConfigurationLoader class to use println instead of LOG.
info.

It is possible to instantiate the IndexingConfiguration without
parameters. This class now extends the ConfigurationLoader class
so the general methods like get or set properties can be done using
reflection.